+++ source/Transform.cpp
@@ -1,52 +1,55 @@
 #include "Transform.h"
+
 #include "ConstantValues.h"
+#include "Matrix3D.h"
+#include "Vector3D.h"
 
 namespace pxd {
 Matrix3D
-MakeRotationX(float angle)
+MakeRotationX(double angle)
 {
   double rad    = DegToRad(angle);
-  float  cosine = cos(rad);
-  float  sine   = sin(rad);
+  double cosine = cos(rad);
+  double sine   = sin(rad);
 
   return Matrix3D(1.0f, 0.0f, 0.0f, 0.0f, cosine, -sine, 0.0f, sine, cosine);
 }
 
 Matrix3D
-MakeRotationY(float angle)
+MakeRotationY(double angle)
 {
   double rad    = DegToRad(angle);
-  float  cosine = cos(rad);
-  float  sine   = sin(rad);
+  double cosine = cos(rad);
+  double sine   = sin(rad);
 
   return Matrix3D(cosine, 0.0f, sine, 0.0f, 1.0f, 0.0f, -sine, 0.0f, cosine);
 }
 
 Matrix3D
-MakeRotationZ(float angle)
+MakeRotationZ(double angle)
 {
   double rad    = DegToRad(angle);
-  float  cosine = cos(rad);
-  float  sine   = sin(rad);
+  double cosine = cos(rad);
+  double sine   = sin(rad);
 
   return Matrix3D(cosine, -sine, 0.0f, sine, cosine, 0.0f, 0.0f, 0.0f, 1.0f);
 }
 
 Matrix3D
-MakeRotation(const Vector3D& axis, float angle)
+MakeRotation(const Vector3D& axis, double angle)
 {
   double rad       = DegToRad(angle);
-  float  sine      = sin(rad);
-  float  cosine    = cos(rad);
-  float  oneMinCos = 1.0f - cosine;
+  double sine      = sin(rad);
+  double cosine    = cos(rad);
+  double oneMinCos = 1.0f - cosine;
 
-  float x = axis.x * oneMinCos;
-  float y = axis.y * oneMinCos;
-  float z = axis.z * oneMinCos;
+  double x = axis.x * oneMinCos;
+  double y = axis.y * oneMinCos;
+  double z = axis.z * oneMinCos;
 
-  float axay = x * axis.y;
-  float axaz = x * axis.z;
-  float ayaz = y * axis.z;
+  double axay = x * axis.y;
+  double axaz = x * axis.z;
+  double ayaz = y * axis.z;
 
   return Matrix3D(cosine + x * axis.x,
                   axay - sine * axis.z,
@@ -62,13 +65,13 @@ MakeRotation(const Vector3D& axis, float angle)
 Matrix3D
 MakeReflection(const Vector3D& axis)
 {
-  float x = axis.x * -2.0f;
-  float y = axis.y * -2.0f;
-  float z = axis.z * -2.0f;
+  double x = axis.x * -2.0f;
+  double y = axis.y * -2.0f;
+  double z = axis.z * -2.0f;
 
-  float axay = x * axis.y;
-  float axaz = x * axis.z;
-  float ayaz = y * axis.z;
+  double axay = x * axis.y;
+  double axaz = x * axis.z;
+  double ayaz = y * axis.z;
 
   return Matrix3D(x * axis.x + 1.0f,
                   axay,
@@ -84,13 +87,13 @@ MakeReflection(const Vector3D& axis)
 Matrix3D
 MakeInvolution(const Vector3D& axis)
 {
-  float x = axis.x * 2.0f;
-  float y = axis.y * 2.0f;
-  float z = axis.z * 2.0f;
+  double x = axis.x * 2.0f;
+  double y = axis.y * 2.0f;
+  double z = axis.z * 2.0f;
 
-  float axay = x * axis.y;
-  float axaz = x * axis.z;
-  float ayaz = y * axis.z;
+  double axay = x * axis.y;
+  double axaz = x * axis.z;
+  double ayaz = y * axis.z;
 
   return Matrix3D(x * axis.x - 1.0f,
                   axay,
@@ -104,23 +107,23 @@ MakeInvolution(const Vector3D& axis)
 }
 
 Matrix3D
-MakeScale(float sx, float sy, float sz)
+MakeScale(double sx, double sy, double sz)
 {
   return Matrix3D(sx, 0.0f, 0.0f, 0.0f, sy, 0.0f, 0.0f, 0.0f, sz);
 }
 
 Matrix3D
-MakeScale(float scale, const Vector3D& v)
+MakeScale(double scale, const Vector3D& v)
 {
   scale -= 1.0f;
 
-  float x = v.x * scale;
-  float y = v.y * scale;
-  float z = v.z * scale;
+  double x = v.x * scale;
+  double y = v.y * scale;
+  double z = v.z * scale;
 
-  float axay = x * v.y;
-  float axaz = x * v.z;
-  float ayaz = y * v.z;
+  double axay = x * v.y;
+  double axaz = x * v.z;
+  double ayaz = y * v.z;
 
   return Matrix3D(x * v.x + 1.0f,
                   axay,
@@ -133,12 +136,12 @@ MakeScale(float scale, const Vector3D& v)
                   z * v.z + 1.0f);
 }
 Matrix3D
-MakeSkew(float angle, const Vector3D& dir, const Vector3D& height)
+MakeSkew(double angle, const Vector3D& dir, const Vector3D& height)
 {
-  angle   = std::tan(angle);
-  float x = dir.x * angle;
-  float y = dir.y * angle;
-  float z = dir.z * angle;
+  angle    = std::tan(angle);
+  double x = dir.x * angle;
+  double y = dir.y * angle;
+  double z = dir.z * angle;
 
   return Matrix3D(x * height.x + 1.0f,
                   x * height.y,
